# openapi.yaml
openapi: 3.0.3
info:
  title: MediaHub API
  description: |
    Contrato mínimo gerado a partir de `server/server.mjs`.
    Autenticação: header `Authorization: Bearer <JWT>` quando indicado.
    Alguns endpoints de busca aceitam token opcional (chaves do provedor por usuário).
  version: "1.0.0"
servers:
  - url: http://localhost:8081
    description: API local (padrão `PORT` no .env)
  - url: https://api.example.com
    description: Substituir pela URL real em produção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      summary: Registro
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Sucesso
  /api/auth/login:
    post:
      tags: [auth]
      summary: Login
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Token / usuário
...
  /api/football/schedule:
    get:
      tags: [football]
      summary: Jogos do dia (premium/admin)
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
          description: YYYY-MM-DD opcional
      responses:
        "200":
          description: date, updatedAt, matches[]
  /api/football/crest:
    get:
      tags: [football]
      summary: Proxy de escudo (imagem)
      description: Requer autenticação premium/admin.
      parameters:
        - in: query
          name: url
          required: true
          schema:
            type: string
            maxLength: 3000
      responses:
        "200":
          description: Imagem
        "401":
          description: Não autenticado
        "403":
          description: Sem plano elegível
    post:
      tags: [football]
      summary: Proxy de escudo via POST (URLs longas)
      description: Requer autenticação premium/admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  maxLength: 16384
      responses:
        "200":
          description: Imagem
        "401":
          description: Não autenticado
        "403":
          description: Sem plano elegível
  /api/football/schedule/refresh:
    get:
      tags: [football]
      summary: Dispara/atualiza coleta (premium/admin)
      description: Requer autenticação premium/admin. Admin também tem POST /api/admin/football/refresh.
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
        "401":
          description: Não autenticado
        "403":
          description: Sem plano elegível
...